requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
anthropic>=0.40,<1
httpx>=0.27,<1
apscheduler>=3.10,<4
pyyaml>=6,<7
python-dotenv>=1,<2
jinja2>=3.1,<4
